dalle_2/requirements.txt
dalle2-pytorch
open-clip-torch